+++ bricks/riverpod_simple_architecture/__brick__/lib/core/router/router.dart
@@ -3,7 +3,7 @@ import 'package:{{project_name.snakeCase()}}/core/router/router.gr.dart';
 
 /// This class used for defined routes and paths na dother properties
 @AutoRouterConfig()
-class AppRouter extends $AppRouter {
+class AppRouter extends RootStackRouter {
   @override
   late final List<AutoRoute> routes = [
     AutoRoute(