+++ bricks/riverpod_simple_architecture/__brick__/lib/core/router/router.gr.dart
@@ -12,23 +12,6 @@ import 'package:auto_route/auto_route.dart' as _i2;
 import 'package:{{project_name.snakeCase()}}/features/counter/view/counter_page.dart'
     deferred as _i1;
 
-abstract class $AppRouter extends _i2.RootStackRouter {
-  $AppRouter({super.navigatorKey});
-
-  @override
-  final Map<String, _i2.PageFactory> pagesMap = {
-    CounterRoute.name: (routeData) {
-      return _i2.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i2.DeferredWidget(
-          _i1.loadLibrary,
-          () => _i1.CounterPage(),
-        ),
-      );
-    }
-  };
-}
-
 /// generated route for
 /// [_i1.CounterPage]
 class CounterRoute extends _i2.PageRouteInfo<void> {
@@ -40,5 +23,13 @@ class CounterRoute extends _i2.PageRouteInfo<void> {
 
   static const String name = 'CounterRoute';
 
-  static const _i2.PageInfo<void> page = _i2.PageInfo<void>(name);
+  static _i2.PageInfo page = _i2.PageInfo(
+    name,
+    builder: (data) {
+      return _i2.DeferredWidget(
+        _i1.loadLibrary,
+        () => _i1.CounterPage(),
+      );
+    },
+  );
 }