+++ bricks/riverpod_simple_architecture/__brick__/lib/features/item_data/controller/item_data_pod.dart
@@ -1,9 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:velocity_x/velocity_x.dart';
-
-final itemDataPod =
-    FutureProvider.autoDispose.family<int, int>((ref, input) async {
-  await Future.delayed(3.seconds);
-  ref.keepAlive();
-  return input;
-});