+++ bricks/riverpod_simple_architecture/__brick__/lib/features/theme_segmented_btn/view/theme_segmented_btn.dart
@@ -37,7 +37,7 @@ class _ThemeSegmentedBtnState extends ConsumerState<ThemeSegmentedBtn> {
             .changeTheme(thememodes.first);
       },
       style: const ButtonStyle(
-          maximumSize: WidgetStatePropertyAll(Size.fromWidth(12))),
+          maximumSize: MaterialStatePropertyAll(Size.fromWidth(12))),
     );
   }
 }