+++ bricks/riverpod_simple_architecture/__brick__/test/app/view/app_test.dart
@@ -24,7 +24,9 @@ Future<void> main() async {
       await tester.pumpApp(
         ProviderScope(
           overrides: [
-            enableInternetCheckerPod.overrideWithValue(false),
+            enableInternetCheckerPod.overrideWith(
+              (ref) => false,
+            ),
             appBoxProvider.overrideWithValue(appBox),
           ],
           child: const CounterPage(),