+++ bricks/riverpod_simple_architecture/__brick__/test/shared/pods/internet_checker_pod_test.dart
@@ -4,6 +4,7 @@ import 'package:internet_connection_checker_plus/internet_connection_checker_plu
 
 import 'package:{{project_name.snakeCase()}}/shared/pods/internet_checker_pod.dart';
 
+import '../widgets/no_interenet_widget_test.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -15,12 +16,16 @@ void main() {
         () async {
           final container = ProviderContainer(
             overrides: [
-              enableInternetCheckerPod.overrideWithValue(false),
+              enableInternetCheckerPod.overrideWithValue(
+                false,
+              ),
             ],
           );
+
           addTearDown(container.dispose);
 
-          final status = await container.read(internetCheckerPod.future);
+          final status =
+              await container.read(internetCheckerNotifierPod.future);
           expect(status, InternetStatus.connected);
         },
       );
@@ -28,29 +33,38 @@ void main() {
         'if internt checker enabled then by deafult internetstatus should be connected',
         () async {
           final container = ProviderContainer(
-            overrides: [enableInternetCheckerPod.overrideWithValue(false)],
+            overrides: [
+              enableInternetCheckerPod.overrideWithValue(
+                true,
+              ),
+            ],
           );
+
           addTearDown(container.dispose);
 
-          final status = await container.read(internetCheckerPod.future);
+          var status = await container.read(internetCheckerNotifierPod.future);
           expectLater(
-            await container
-                .read(internetConnectionCheckerPod)
-                .hasInternetAccess,
-            equals(true),
+            status,
+            equals(InternetStatus.connected),
           );
+
           container
-              .read(internetConnectionCheckerPod)
+              .read(internetConnectionCheckerInstancePod)
               .onStatusChange
               .listen((event) {});
+          final newStatus =
+              await container.read(internetCheckerNotifierPod.future);
 
-          expect(status, InternetStatus.connected);
-          expect(container.read(internetConnectionCheckerPod).checkInterval,
+          expect(newStatus, InternetStatus.connected);
+          expect(
+              container
+                  .read(internetConnectionCheckerInstancePod)
+                  .checkInterval,
               equals(const Duration(seconds: 5)));
           container
-              .read(internetCheckerPod.notifier)
+              .read(internetCheckerNotifierPod.notifier)
               .change(status: InternetStatus.disconnected);
-          expect(container.read(internetCheckerPod).requireValue,
+          expect(container.read(internetCheckerNotifierPod).requireValue,
               equals(InternetStatus.disconnected));
         },
       );
@@ -58,29 +72,33 @@ void main() {
         'if internet checker enabled then by deafult internetstatus should be connected',
         () async {
           final container = ProviderContainer(
-            overrides: [],
+            overrides: [
+              enableInternetCheckerPod.overrideWithValue(
+                true,
+              ),
+            ],
           );
           addTearDown(container.dispose);
 
-          final status = await container.read(internetCheckerPod.future);
+          var status = await container.read(internetCheckerNotifierPod.future);
           expectLater(
-            await container
-                .read(internetConnectionCheckerPod)
-                .hasInternetAccess,
-            equals(true),
+            status,
+            equals(InternetStatus.connected),
           );
           container
-              .read(internetConnectionCheckerPod)
+              .read(internetConnectionCheckerInstancePod)
               .onStatusChange
               .listen((event) {});
 
-          expect(status, InternetStatus.connected);
-          expect(container.read(internetConnectionCheckerPod).checkInterval,
+          expect(
+              container
+                  .read(internetConnectionCheckerInstancePod)
+                  .checkInterval,
               equals(const Duration(seconds: 5)));
           container
-              .read(internetCheckerPod.notifier)
+              .read(internetCheckerNotifierPod.notifier)
               .change(status: InternetStatus.disconnected);
-          expect(container.read(internetCheckerPod).requireValue,
+          expect(container.read(internetCheckerNotifierPod).requireValue,
               equals(InternetStatus.disconnected));
         },
       );